remove compilation warnings due to incompatible types conversions on int/long variables

+++ include/zpoller.h
@@ -62,7 +62,7 @@ CZMQ_EXPORT void
 //  calling zpoller_expired () and zpoller_terminated (). The timeout is in
 //  msec.
 CZMQ_EXPORT void *
-    zpoller_wait (zpoller_t *self, int timeout);
+    zpoller_wait (zpoller_t *self, long timeout);
 
 //  Return true if the last zpoller_wait () call ended because the timeout
 //  expired, without any error.
